Update ge-jasco-zwave-plus-on-off-switch.groovy
Added support for double tap and hold up or down for button 1 and 2.  Updated to make sure button values and number of buttons are updated.  If "held" button value does not appear for existing installs, change any preference in device settings (can then immediately be changed back) to add the "held" button value.