Update ge-jasco-zwave-plus-dimmer-switch.groovy
Code added based on Even Hunter's suggestions to add configuration paremeters that were not on the zwave alliance's database, which include the ability to make the switch an on/off switch and set a minimum dimmer level.  Other various code cleanup, added code to make sure button number/value is updated, and ability to force settings reset.